sched: fix issue of cpu freq running at max always
All cpus are running at max freq, the reason is that in the check of
sugov_up_down_rate_limit() in cpufreq_schedutil.c, the time passed in
is always 0, so the check is always true and hence the freq will not
be updated. It is caused by sched_ktime_clock() will return 0 if
CONFIG_SCHED_WALT is not set. Fix it by replacing sched_ktime_clock()
with ktime_get_ns().

Bug: 119932718
Test: cpu freq could change after fix
Change-Id: I62a0b35208dcd7a1d23da27f909cce3e59208d1f
Signed-off-by: Rick Yiu <rickyiu@google.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: twisted <36546624+TwistedPrime@users.noreply.github.com>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>